requirements.txt
discord==1.0.1
discord.py==1.5.1
python-dotenv==0.15.0
requests==2.24.0

